requirements.txt
fastapi
uvicorn

pydantic
python-dotenv
loguru

telethon
